Fix errors selection
