+++ tests/utils.py
@@ -1,0 +1,14 @@
+import os
+from contextlib import contextmanager
+
+
+@contextmanager
+def chdir(path):
+    """Context manager for changing dir and restoring previous workdir after exit.
+    """
+    curdir = os.getcwd()
+    os.chdir(str(path))
+    try:
+        yield
+    finally:
+        os.chdir(curdir)
